+++ CHANGELOG.md
@@ -8,12 +8,22 @@ Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 ## [Unreleased]
 #### Added
-- Trim for input field 
+- Trim for input field
+
+#### Changed
+- Moved logo and favicon to branding folder
+
+### "Under the hood" changes
+#### Added
+- `APP_NAME` in config
 
 ## [1.10.3-rc.1] - 2019-10-18
 #### Fixed
 - A bug when the user can not enter a value in input-field with type 'number' on Firefox
 - A bug with infinite loading of the page "Register of shares"
+
+#### Changed
+- Color scheme improvements
 
 ## [1.10.3-rc.0] - 2019-10-11
 #### Added